requirements.txt
flask==3.0.0
requests==2.31.0
opencv-python==4.8.1.78
scikit-image==0.22.0
pillow==10.1.0
pytest==7.4.3